renamed components, functions and variables to be more explicit
